Port build and Python bindings to Qt6/PyQt6 and modernize build settings
Switch project to Qt6 (find Qt6 components, Qt6Keychain, PyQt6) and update CMake/packaging for Qt6 plugins and tools. Bump C++ standard to C++20 and adapt CMake find-modules (add QScintilla/Qwt helpers, update QCA/QCATOOL lookups), move Python SIP/packaging layout to PyQt6, and remove obsolete PyQt5/HDF5 helpers.

Also apply small code and binding adjustments required by the port: use QtConcurrent lambda, handle container/size type conversions, use QTimeZone::UTC, add includes (QElapsedTimer, QDateTime), and various autogenerated SIP tweaks to match Qt6/PyQt6 bindings.

These changes prepare the codebase and build system for Qt6/PyQt6 usage and modern C++ compilation.

+++ src/core/GIS/private/reosmesh_p.cpp
@@ -1920,7 +1920,7 @@ QgsMeshDataBlock ReosResultDataset::areFacesActive( int faceIndex, int count ) c
 
   const QVector<int> &sourceValues = mSimulationResult->activeFaces( mDatasetIndex );
 
-  int valueCount = std::min( count, sourceValues.size() - faceIndex );
+  int valueCount = std::min( count, static_cast<int>( sourceValues.size() ) - faceIndex );
   QVector<int> values( valueCount );
   memcpy( values.data(), &( sourceValues[faceIndex] ), count * sizeof( int ) );
   QgsMeshDataBlock ret( QgsMeshDataBlock::ActiveFlagInteger, valueCount );
